Hacked together Approve and Deny
It's not the most secure, but it'll work for the purposes of our presentation.  Works on the backend, all is checked on my end.

Getting stuff to carry between a page accessed by GET and accessed by POST is whack.  References included in code for later.  We'll get it in cleanup.

Co-Authored-By: Luke <55993047+Luke-El@users.noreply.github.com>